Moved check_warnings decorator in utils for resusability purpose
